Update events_handler.js
Updated tabs.remove logic to handle special case when 'add url' button is clicked on the Pinboard.in site itself.

In previous versions the tab would erroneously be closed. Now, we check for the 'source' param and only close the tab if it = quick_pin. Otherwise, get the value of the 'next' param (used on Pinboard.in) and redirect the user to that instead which is the expected behavior.

- getUrlVar callback borrowed from https://gist.github.com/1771618